Relanding: Split content_shell dependencies for layout tests.
Reason for relanding:
The Blink bot that failed was using an untested build config,
which has been fixed in crrev.com/224733008.

Original issue's description:
> Split content_shell dependencies for layout tests.
>
> This CL is isolating the minimum dependencies required to make layout
> tests work in content_shell without requiring it to depend on the entire
> test_support_content_target.
> crrev.com/11275277 has introduced a dependency on test_support_content
> to cope with layout test. test_support_content, however, was intended
> for unit/browser-tests and adding its dependency to content_shell
> causes a lot of unnecessary sources to be linked in.
> The linker --gc-sections has hidden the problem so far, but that doesn't
> seem to be an option for arm64 (at least with the current toolchain).


BUG=159847, 354405
TBR=jam@chromium.org

Review URL: https://codereview.chromium.org/225783005

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@262120 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 5326cd2ffbb54a1ac573e19a233b77b45ab9b44c

+++ content/test/mock_webclipboard_impl.cc
@@ -102,7 +102,7 @@ blink::WebData MockWebClipboardImpl::readImage(
                         gfx::PNGCodec::FORMAT_BGRA,
 #endif
                         gfx::Size(bitmap.width(), bitmap.height()),
-                        bitmap.rowBytes(),
+                        static_cast<int>(bitmap.rowBytes()),
                         false /* discard_transparency */,
                         std::vector<gfx::PNGCodec::Comment>(),
                         &encoded_image);
